Make Block movable, align on size_t for size properties.
* Use size_t for dimensional properties to align with rest of code
  (and general container `.size()` types)
* Adopt MonoView for views
* Remove `Swap()` and `SetNumChannels()` methods to emphasize the
  implicit constness of the class (i.e. encourage setting dimensions
  only at construction time).

Bug: chromium:335805780

+++ modules/audio_processing/aec3/echo_remover_unittest.cc
@@ -156,8 +156,8 @@ TEST(EchoRemover, BasicEchoRemoval) {
           delay_buffers[band].resize(x.NumChannels());
         }
 
-        for (int band = 0; band < x.NumBands(); ++band) {
-          for (int channel = 0; channel < x.NumChannels(); ++channel) {
+        for (size_t band = 0; band < x.NumBands(); ++band) {
+          for (size_t channel = 0; channel < x.NumChannels(); ++channel) {
             delay_buffers[band][channel].reset(
                 new DelayBuffer<float>(delay_samples));
           }
@@ -168,8 +168,8 @@ TEST(EchoRemover, BasicEchoRemoval) {
         for (int k = 0; k < kNumBlocksToProcess; ++k) {
           const bool silence = k < 100 || (k % 100 >= 10);
 
-          for (int band = 0; band < x.NumBands(); ++band) {
-            for (int channel = 0; channel < x.NumChannels(); ++channel) {
+          for (size_t band = 0; band < x.NumBands(); ++band) {
+            for (size_t channel = 0; channel < x.NumChannels(); ++channel) {
               if (silence) {
                 std::fill(x.begin(band, channel), x.end(band, channel), 0.f);
               } else {
